feat: set app user document when custom token signin
