+++ functions/src/index.ts
@@ -28,7 +28,7 @@ admin.initializeApp({
 functions.setGlobalOptions({ region: `asia-northeast1` })
 
 /** ここでデプロイする関数をまとめる。 */
-import { createfirebaseauthcustomtoken } from './callable-functions/createFirebaseAuthCustomToken'
+import { createfirebaseauthcustomtoken } from './createFirebaseAuthCustomToken'
 
 /** index.ts で import してデプロイする関数一覧。 */
 export { createfirebaseauthcustomtoken }
